Move logic bc cool and other stuff
